Lodders09 and NIST data dictionaries added
Created all dictionaries in `data` folder for Lodders+ 2009 and NIST
isotope ratios. These also contain tests to ensure that the data
stays the same (in case of recalculation, the tests need to be adopted).

Rewrote the initialization of main such that the correct element and
isotope dictionaries are automatically loaded into private variables.
Also adjusted the test.

The good news: There is not data file required anymore and no file
reading and writing. The parsers are stored in the `dev` folder on
github and are NOT part of the package or the tests. The tests check
the package and check for data integrity, thus no test for dev is
required. Also: If the data change, these parsers probably need to be
slightly adjusted.