ENH: Better implementation of `abundance_unit` routine.
Abundance unit routine now simply sets a bool `self._log_abu` value.
Logarithmic abundance dictionaries are already created when database
is initialized, allowing that it does not have to be recalculated
on a switch.
*args, **kwargs are now passed through the proxy class and are accepted
by the Isotopes and Elements classes. This allows to directly pass the
`self._log_abu` bool on without any issue.
If (currently not possible) one of those classes is initialized without
a `log_abu` value, it is generated in the __init__ and set by default to
False.